GALAXY-107: delete user capabilities
Fix problem whereby you could have workspaces and artifacts with the same
name inside a workspace.
(Hopefully this fixes the concurrency issue as well)

Return a 409 if the workspace/artifact already exists.

Add a logging.properties so the web-standalone spits out more logging
(Need to still fully test this in a bit. I don't understand this logging
mess we've got ourselves into in Java so if this is wrong let me know).





git-svn-id: http://svn.muleforge.org/galaxy/branches/galaxy-1.0-beta-2@459 a180d455-963b-0410-bd2a-f8bbb34fb4c7