Link property editing.
I changed Extension so it had full control over
what exactly is stored. This way if you have information
which is stored in a separate JCR tree (e.g. Links), it 
does not have to be duplicated inside the Item's properties.

Only bug which I'm aware of at this point is that Links
are not checked for duplicates. Working on this. 


git-svn-id: http://svn.muleforge.org/galaxy/trunk@1289 a180d455-963b-0410-bd2a-f8bbb34fb4c7